Refactor admin routes.
