+++ src/app/(root)/(dashboard)/workspace/[slug]/admin/recordings/page.tsx
