+++ src/app/(root)/(dashboard)/workspace/[slug]/admin/workspace-settings/_components/danger-zone.tsx
